[example] make gpt example directory more clear
