--- examples/language/gpt/auto_parallel_with_gpt/requirements.txt
@@ -1,4 +1,0 @@
-colossalai >= 0.1.12
-torch >= 1.8.1
-transformers >= 4.231
-PuLP >= 2.7.0
