+++ examples/language/gpt/gemini/train_gpt_demo.py
@@ -5,9 +5,10 @@ from time import time
 import psutil
 import torch
 import torch.nn as nn
+from commons.model_zoo import model_builder
+from commons.utils import get_data, get_tflops
 from packaging import version
 from torch.nn.parallel import DistributedDataParallel as DDP
-from utils import get_data, get_tflops
 
 import colossalai
 from colossalai.logging import disable_existing_loggers, get_dist_logger
@@ -15,7 +16,6 @@ from colossalai.nn.parallel import ZeroDDP
 from colossalai.tensor import ColoParameter, ComputePattern, ComputeSpec, ProcessGroup, ReplicaSpec, ShardSpec
 from colossalai.utils import get_current_device
 from colossalai.utils.model.colo_init_context import ColoInitContext
-from model_zoo import model_builder
 
 CAI_VERSION = colossalai.__version__
